server/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
astroid==2.4.2
certifi==2020.12.5
colorama==0.4.4
Django @ file:///tmp/build/80754af9/django_1613762073373/work
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
djangorestframework==3.12.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pylint==2.6.0
python-decouple==3.4
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
toml==0.10.2
wincertstore==0.2
wrapt==1.12.1
